Removes obsolete AasUtils methods and tidies warnings
