+++ dataformat-core/src/main/java/org/eclipse/digitaltwin/aas4j/v3/dataformat/core/serialization/EnumSerializer.java
@@ -17,7 +17,6 @@ package org.eclipse.digitaltwin.aas4j.v3.dataformat.core.serialization;
 
 import java.io.IOException;
 
-import org.eclipse.digitaltwin.aas4j.v3.dataformat.core.util.AasUtils;
 import org.eclipse.digitaltwin.aas4j.v3.dataformat.core.util.ReflectionHelper;
 import org.eclipse.digitaltwin.aas4j.v3.model.DataTypeDefXSD;
 import org.eclipse.digitaltwin.aas4j.v3.model.DataTypeIec61360;
@@ -33,12 +32,11 @@ import com.fasterxml.jackson.databind.SerializerProvider;
  * be converted from SCREAMING_SNAKE_CASE to UpperCamelCase, else default
  * serialization will be used
  */
+@SuppressWarnings("rawtypes")
 public class EnumSerializer extends JsonSerializer<Enum> {
 
-    protected static final char UNDERSCORE = '_';
-
-    @Override
-    public void serialize(Enum value, JsonGenerator gen, SerializerProvider provider) throws IOException {
+	@Override
+	public void serialize(Enum value, JsonGenerator gen, SerializerProvider provider) throws IOException {
         if (value instanceof DataTypeDefXSD) {
             // only for the DataTypeDefXsd notation
             if (value.equals(DataTypeDefXSD.ANY_URI)) {
@@ -49,7 +47,7 @@ public class EnumSerializer extends JsonSerializer<Enum> {
 				handleTimeRelatedValue(gen, value);
             } else {
                 // pattern: 'xs:' + camelCase
-                String enum_string = AasUtils.serializeEnumName(value.name());
+				String enum_string = serializeEnumName(value.name());
                 gen.writeString("xs:" + enum_string.substring(0, 1).toLowerCase() + enum_string.substring(1));
             }
         } else if (value instanceof DataTypeIec61360) {
@@ -57,21 +55,43 @@ public class EnumSerializer extends JsonSerializer<Enum> {
         } else if (value instanceof Direction || value instanceof StateOfEvent) {
             gen.writeString(value.name().toLowerCase());
         } else if (ReflectionHelper.ENUMS.contains(value.getClass())) {
-            gen.writeString(AasUtils.serializeEnumName(value.name()));
+			gen.writeString(serializeEnumName(value.name()));
         } else {
             provider.findValueSerializer(Enum.class).serialize(value, gen, provider);
         }
     }
 
-	private void handleTimeRelatedValue(JsonGenerator gen, Enum value) throws IOException {
-		String enum_string = AasUtils.serializeEnumName(value.name());
+	private void handleTimeRelatedValue(JsonGenerator gen, Enum<?> value) throws IOException {
+		String enum_string = serializeEnumName(value.name());
 		String adaptedEnumString = "xs:g" + enum_string.substring(1, 2).toUpperCase() + enum_string.substring(2);
 		gen.writeString(adaptedEnumString);
 	}
 
-	private boolean isTimeRelatedValue(Enum value) {
-		String enum_string = AasUtils.serializeEnumName(value.name());
+	private boolean isTimeRelatedValue(Enum<?> value) {
+		String enum_string = serializeEnumName(value.name());
 		return enum_string.startsWith("G");
+	}
+
+	/**
+	 * Translates an enum value from SCREAMING_SNAKE_CASE to CamelCase
+	 *
+	 * @param input
+	 *            input name in SCREAMING_SNAKE_CASE
+	 * @return name in CamelCase
+	 */
+	public static String serializeEnumName(String input) {
+		String result = "";
+		boolean capitalize = true;
+		for (int i = 0; i < input.length(); i++) {
+			char currentChar = input.charAt(i);
+			if ('_' == currentChar) {
+				capitalize = true;
+			} else {
+				result += capitalize ? currentChar : Character.toLowerCase(currentChar);
+				capitalize = false;
+			}
+		}
+		return result;
 	}
 
 }
