+++ dataformat-core/src/main/java/org/eclipse/digitaltwin/aas4j/v3/dataformat/core/util/AasUtils.java
@@ -15,22 +15,35 @@
  */
 package org.eclipse.digitaltwin.aas4j.v3.dataformat.core.util;
 
-import com.google.common.reflect.TypeToken;
-import org.eclipse.digitaltwin.aas4j.v3.model.*;
-import org.eclipse.digitaltwin.aas4j.v3.model.impl.DefaultKey;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import java.beans.IntrospectionException;
 import java.beans.Introspector;
 import java.beans.PropertyDescriptor;
 import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Type;
-import java.util.*;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Comparator;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Optional;
+import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+
+import org.eclipse.digitaltwin.aas4j.v3.dataformat.core.deserialization.EnumDeserializer;
+import org.eclipse.digitaltwin.aas4j.v3.dataformat.core.serialization.EnumSerializer;
+import org.eclipse.digitaltwin.aas4j.v3.model.Environment;
+import org.eclipse.digitaltwin.aas4j.v3.model.Identifiable;
+import org.eclipse.digitaltwin.aas4j.v3.model.Key;
+import org.eclipse.digitaltwin.aas4j.v3.model.KeyTypes;
+import org.eclipse.digitaltwin.aas4j.v3.model.Operation;
+import org.eclipse.digitaltwin.aas4j.v3.model.Referable;
+import org.eclipse.digitaltwin.aas4j.v3.model.Reference;
+import org.eclipse.digitaltwin.aas4j.v3.model.ReferenceTypes;
+import org.eclipse.digitaltwin.aas4j.v3.model.SubmodelElementList;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import com.google.common.reflect.TypeToken;
 
 /**
  * Provides utility functions related to AAS
@@ -39,223 +52,26 @@ public class AasUtils {
 
     private static final Logger log = LoggerFactory.getLogger(AasUtils.class);
 
-    private static final char UNDERSCORE = '_';
-    private static final String KEY_REGEX_GROUP_TYPE = "type";
-    private static final String KEY_REGEX_GROUP_VALUE = "value";
-    private static final Pattern KEY_REGEX = Pattern.compile(
-            String.format("\\((?<%s>\\w+)\\)(?<%s>.*)",
-                    KEY_REGEX_GROUP_TYPE,
-                    KEY_REGEX_GROUP_VALUE));
-    private static final String REFERENCE_REGEX_GROUP_TYPE = "value";
-    private static final Pattern REFERENCE_TYPE_PREFIX_REGEX = Pattern.compile(
-            String.format("\\[(?<%s>[a-zA-Z]+)\\].*", REFERENCE_REGEX_GROUP_TYPE));
-
-    private static final String REFERENCE_ELEMENT_DELIMITER = ", ";
+	private static final String REFERENCE_ELEMENT_DELIMITER = ", ";
 
     private AasUtils() {
     }
 
-    /**
-     * Formats a Reference as string
-     *
-     * @param reference Reference to serialize
-     * @return string representation of the reference for serialization, null if reference is null
-     */
-    public static String asString(Reference reference) {
-        if (reference == null) {
-            return null;
-        }
-        return String.format("[%s]%s", reference.getType(),
-                reference.getKeys().stream()
-                        .map(x -> String.format("(%s)%s",
-                                serializeEnumName(x.getType().name()),
-                                x.getValue()))
-                        .collect(Collectors.joining(REFERENCE_ELEMENT_DELIMITER)));
-    }
-
-    /**
-     * Parses a given string as Reference. If the given string is not a valid reference, null is returned.
-     *
-     * @param value String representation of the reference
-     * @return parsed Reference or null is given value is not a valid Reference
-     */
-    public static Reference parseReference(String value) {
-        return parseReference(value,
-                ReflectionHelper.getDefaultImplementation(Reference.class),
-                ReflectionHelper.getDefaultImplementation(Key.class));
-    }
-
-    /**
-     * Parses a given string as Reference using the provided implementation of Reference and Key interface. If the given
-     * string is not a valid reference, null is returned.
-     *
-     * @param value String representation of the reference
-     * @param referenceType implementation type of Reference interface
-     * @param keyType implementation type of Key interface
-     * @return parsed Reference or null is given value is not a valid Reference
-     */
-    public static Reference parseReference(String value, Class<? extends Reference> referenceType, Class<? extends Key> keyType) {
-        String reference = value;
-        if (reference == null || reference.isBlank()) {
-            return null;
-        }
-        try {
-            Reference result = referenceType.getConstructor().newInstance();
-            // check if optional [<ReferenceTypes>] is present, if so, check for consistency
-            Matcher matcher = REFERENCE_TYPE_PREFIX_REGEX.matcher(reference);
-            if (matcher.find()) {
-                result.setType(ReferenceTypes.valueOf(deserializeEnumName(matcher.group(REFERENCE_REGEX_GROUP_TYPE))));
-                reference = reference.substring(reference.indexOf(serializeEnumName(result.getType().name())) + 1);
-
-            }
-            result.setKeys(Stream.of(reference.split(REFERENCE_ELEMENT_DELIMITER))
-                    .map(x -> parseKey(x))
-                    .collect(Collectors.toList()));
-            if (!result.getKeys().isEmpty()) {
-                if (result.getType() == null) {
-                    // deduct from first element
-                    result.setType(result.getKeys().get(0).getType() == KeyTypes.GLOBAL_REFERENCE
-                            ? ReferenceTypes.EXTERNAL_REFERENCE
-                            : ReferenceTypes.MODEL_REFERENCE);
-                } else {
-                    // validate against first element
-                    if (!isCompatible(result.getKeys().get(0).getType(), result.getType())) {
-                        throw new IllegalArgumentException(String.format("invalid reference - reference type '%s' is not compatible with type of first key elemenet '%s'",
-                                result.getType(), result.getKeys().get(0)));
-                    }
-                }
-            }
-            // check that keys following SubmodelElementList have valid index (i.e. are number >= 0)
-            validateSubmodelElementListKeyValues(result.getKeys());
-            return result;
-        } catch (NoSuchMethodException | SecurityException | InstantiationException | IllegalAccessException | IllegalArgumentException | InvocationTargetException ex) {
-            throw new IllegalArgumentException("error parsing reference - could not instantiate reference type", ex);
-        }
-    }
-
-    private static void validateSubmodelElementListKeyValues(List<Key> keys) {
-        if (keys == null || keys.size() <= 1) {
-            return;
-        }
-        for (int i = 0; i < keys.size() - 1; i++) {
-            if (keys.get(i).getType() == KeyTypes.SUBMODEL_ELEMENT_LIST) {
-                try {
-                    if (Integer.parseInt(keys.get(i + 1).getValue()) < 0) {
-                        throw new IllegalArgumentException(String.format("invalid value for key with index %d, expected integer values >= 0, but found '%s'",
-                                i, keys.get(i + 1).getValue()));
-                    }
-                } catch (NumberFormatException ex) {
-                    throw new IllegalArgumentException(String.format("invalid value for key with index %d, expected integer values >= 0, but found '%s'",
-                            i, keys.get(i + 1).getValue()));
-                }
-            }
-        }
-    }
-
-    private static boolean isCompatible(KeyTypes keyType, ReferenceTypes referenceType) {
-        if (keyType == null && referenceType == null) {
-            return true;
-        }
-        if (keyType == null ^ referenceType == null) {
-            return false;
-        }
-        return referenceType == ReferenceTypes.EXTERNAL_REFERENCE
-                ? keyType == KeyTypes.GLOBAL_REFERENCE
-                : keyType != KeyTypes.GLOBAL_REFERENCE;
-    }
-
-    /**
-     * Gets property with given name as defined in type of given parent or null if not defined
-     *
-     * @param parent parent object
-     * @param propertyName name of the property
-     * @return property with given name as defined in type of given parent or null if not defined
-     */
-    public static PropertyDescriptor getProperty(Object parent, String propertyName) {
-        if (parent == null || propertyName == null || propertyName.isBlank()) {
-            return null;
-        }
-        return getAasProperties(parent.getClass()).stream()
-                .filter(x -> x.getName().equals(propertyName))
-                .findAny()
-                .orElse(null);
-    }
-
-    /**
-     * Gets the content type of a generic collection type
-     *
-     * @param genericCollectionType the generic collection type
-     * @return the content type of the generic collection type
-     */
-    public static Class<?> getCollectionContentType(Type genericCollectionType) {
-        return TypeToken.of(genericCollectionType).resolveType(Collection.class.getTypeParameters()[0]).getRawType();
-    }
-
-    /**
-     * Gets property with given name as defined in given type or null if not defined
-     *
-     * @param type type containing the property
-     * @param propertyName name of the property
-     * @return property with given name as defined in given type or null if not defined
-     */
-    public static PropertyDescriptor getProperty(Class<?> type, String propertyName) {
-        if (type == null || propertyName == null || propertyName.isBlank()) {
-            return null;
-        }
-        return getAasProperties(type).stream()
-                .filter(x -> x.getName().equals(propertyName))
-                .findAny()
-                .orElse(null);
-    }
-
-    /**
-     * Parses a given string as Key. If the given string is not a valid key, null is returned.
-     *
-     * @param value String representation of the key
-     * @return parsed Key or null is given value is not a valid Key
-     */
-    public static Key parseKey(String value) {
-        Matcher matcher = KEY_REGEX.matcher(value);
-        if (matcher.find()) {
-            return new DefaultKey.Builder()
-                    .type(KeyTypes.valueOf(deserializeEnumName(matcher.group(KEY_REGEX_GROUP_TYPE))))
-                    .value(matcher.group(KEY_REGEX_GROUP_VALUE))
-                    .build();
-        }
-        return null;
-    }
-
-    /**
-     * Checks if a reference is a local reference or not. This functionality may not be 100% correct as since v3.0RC01
-     * of the AAS specification there no longer is an isLocal property to check this and no alternative way to determine
-     * whether a reference is local or not is introduced. This method only checks for the presence of any Key with type
-     * GLOBAL_REFERENCE. Another approach would be to actually try resolving the reference locally.
-     *
-     * @param reference The reference to check
-     * @param environment The environment context the reference resides. In current implementation this is not used
-     * @return true if the reference is a local reference to the given environment, false otherwise
-     */
-    public static boolean isLocal(Reference reference, Environment environment) {
-        if (reference == null) {
-            throw new IllegalArgumentException("reference must be non-null");
-        }
-        if (reference.getKeys() == null || reference.getKeys().isEmpty()) {
-            throw new IllegalArgumentException("reference must contain at least one key");
-        }
-        return reference.getKeys().get(0).getType() != KeyTypes.GLOBAL_REFERENCE;
-    }
-
-    public static List<Submodel> getSubmodelTemplates(AssetAdministrationShell aas, Environment environment) {
-        return aas.getSubmodels().stream()
-                .map(ref -> resolve(ref, environment, Submodel.class))
-                .filter(sm -> sm != null)
-                .filter(sm -> sm.getKind() != ModellingKind.INSTANCE)
-                .collect(Collectors.toList());
-    }
-
-    public static boolean hasTemplate(AssetAdministrationShell aas, Environment environment) {
-        return !getSubmodelTemplates(aas, environment).isEmpty();
-    }
+	/**
+	 * Formats a Reference as string
+	 *
+	 * @param reference
+	 *            Reference to serialize
+	 * @return string representation of the reference for serialization, null if
+	 *         reference is null
+	 */
+	private static String asString(Reference reference) {
+		if (reference == null) {
+			return null;
+		}
+		return String.format("[%s]%s", reference.getType(),
+				reference.getKeys().stream().map(x -> String.format("(%s)%s", EnumSerializer.serializeEnumName(x.getType().name()), x.getValue())).collect(Collectors.joining(REFERENCE_ELEMENT_DELIMITER)));
+	}
 
     /**
      * Creates a reference for an Identifiable instance using provided implementation types for reference and key
@@ -299,55 +115,9 @@ public class AasUtils {
     public static KeyTypes referableToKeyType(Referable referable) {
         Class<?> aasInterface = ReflectionHelper.getAasInterface(referable.getClass());
         if (aasInterface != null) {
-            return KeyTypes.valueOf(deserializeEnumName(aasInterface.getSimpleName()));
+            return KeyTypes.valueOf(EnumDeserializer.deserializeEnumName(aasInterface.getSimpleName()));
         }
         return null;
-    }
-
-    /**
-     * Translates an enum value from SCREAMING_SNAKE_CASE to CamelCase
-     *
-     * @param input input name in SCREAMING_SNAKE_CASE
-     * @return name in CamelCase
-     */
-    public static String serializeEnumName(String input) {
-        String result = "";
-        boolean capitalize = true;
-        for (int i = 0; i < input.length(); i++) {
-            char currentChar = input.charAt(i);
-            if (UNDERSCORE == currentChar) {
-                capitalize = true;
-            } else {
-                result += capitalize
-                        ? currentChar
-                        : Character.toLowerCase(currentChar);
-                capitalize = false;
-            }
-        }
-        return result;
-    }
-
-    /**
-     * Translates an enum value from CamelCase to SCREAMING_SNAKE_CASE
-     *
-     * @param input input name in CamelCase
-     * @return name in SCREAMING_SNAKE_CASE
-     */
-    public static String deserializeEnumName(String input) {
-        String result = "";
-        if (input == null || input.isEmpty()) {
-            return result;
-        }
-        result += Character.toUpperCase(input.charAt(0));
-        for (int i = 1; i < input.length(); i++) {
-            char currentChar = input.charAt(i),
-                    previousChar = input.charAt(i - 1);
-            if (Character.isUpperCase(currentChar) && Character.isLowerCase(previousChar)) {
-                result += UNDERSCORE;
-            }
-            result += Character.toUpperCase(currentChar);
-        }
-        return result;
     }
 
     /**
@@ -357,9 +127,9 @@ public class AasUtils {
      * @return a Java interface representing the provided KeyElements type or null if no matching Class/interface could
      * be found. It also returns abstract types like SUBMODEL_ELEMENT or DATA_ELEMENT
      */
-    public static Class<?> keyTypeToClass(KeyTypes key) {
+	private static Class<?> keyTypeToClass(KeyTypes key) {
         return Stream.concat(ReflectionHelper.INTERFACES.stream(), ReflectionHelper.INTERFACES_WITHOUT_DEFAULT_IMPLEMENTATION.stream())
-                .filter(x -> x.getSimpleName().equals(serializeEnumName(key.name())))
+				.filter(x -> x.getSimpleName().equals(EnumSerializer.serializeEnumName(key.name())))
                 .findAny()
                 .orElse(null);
     }
@@ -453,16 +223,6 @@ public class AasUtils {
     }
 
     /**
-     * Creates a deep-copy clone of a reference
-     *
-     * @param reference the reference to clone
-     * @return the cloned reference
-     */
-    public static Reference clone(Reference reference) {
-        return clone(reference, ReflectionHelper.getDefaultImplementation(Reference.class), ReflectionHelper.getDefaultImplementation(Key.class));
-    }
-
-    /**
      * Creates a deep-copy clone of a reference using provided implementation types for reference and key
      *
      * @param reference the reference to clone
@@ -471,7 +231,7 @@ public class AasUtils {
      *
      * @return the cloned reference
      */
-    public static Reference clone(Reference reference, Class<? extends Reference> referenceType, Class<? extends Key> keyType) {
+	private static Reference clone(Reference reference, Class<? extends Reference> referenceType, Class<? extends Key> keyType) {
         if (reference == null || reference.getKeys() == null || reference.getKeys().isEmpty()) {
             return null;
         }
@@ -517,7 +277,8 @@ public class AasUtils {
      * @return returns an instance of T if the reference could successfully be resolved, otherwise null
      * @throws IllegalArgumentException if something goes wrong while resolving
      */
-    public static <T extends Referable> T resolve(Reference reference, Environment env, Class<T> type) {
+	@SuppressWarnings("unchecked")
+	public static <T extends Referable> T resolve(Reference reference, Environment env, Class<T> type) {
         if (reference == null || reference.getKeys() == null || reference.getKeys().isEmpty()) {
             return null;
         }
@@ -579,8 +340,7 @@ public class AasUtils {
                                 key.getValue()));
                     }
                 } else {
-                    Collection collection;
-                    // operation needs special handling because of nested values
+					Collection<?> collection;
                     if (Operation.class.isAssignableFrom(current.getClass())) {
                         Operation operation = (Operation) current;
                         collection = Stream.of(operation.getInputVariables().stream(),
@@ -609,12 +369,12 @@ public class AasUtils {
                                             .collect(Collectors.joining(", "))));
                         }
                         try {
-                            collection = (Collection) matchingProperties.get(0).getReadMethod().invoke(current);
+							collection = (Collection<?>) matchingProperties.get(0).getReadMethod().invoke(current);
                         } catch (IllegalAccessException | IllegalArgumentException | InvocationTargetException ex) {
                             throw new IllegalArgumentException("error resolving reference", ex);
                         }
                     }
-                    Optional next = collection.stream()
+					Optional<?> next = collection.stream()
                             .filter(x -> ((Referable) x).getIdShort().equals(key.getValue()))
                             .findFirst();
                     if (next.isEmpty()) {
@@ -635,7 +395,7 @@ public class AasUtils {
      * @return a list of all properties defined in any of AAS interface implemented by type. If type does not implement
      * any AAS interface an empty list is returned.
      */
-    public static List<PropertyDescriptor> getAasProperties(Class<?> type) {
+	private static List<PropertyDescriptor> getAasProperties(Class<?> type) {
         Class<?> aasType = ReflectionHelper.getAasInterface(type);
         if (aasType == null) {
             aasType = ReflectionHelper.INTERFACES_WITHOUT_DEFAULT_IMPLEMENTATION.stream()
