--- dataformat-core/src/main/java/org/eclipse/digitaltwin/aas4j/v3/dataformat/core/util/MostSpecificClassComparator.java
@@ -1,35 +1,0 @@
-/*
- * Copyright (c) 2021 Fraunhofer-Gesellschaft zur Foerderung der angewandten Forschung e. V.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.eclipse.digitaltwin.aas4j.v3.dataformat.core.util;
-
-import java.util.Comparator;
-
-/**
- * Comparator comparing two classes regarding which type is more specific.
- */
-public class MostSpecificClassComparator implements Comparator<Class<?>> {
-
-    @Override
-    public int compare(Class<?> x, Class<?> y) {
-        if (y.isAssignableFrom(x)) {
-            if (x.isAssignableFrom(y)) {
-                return 0;
-            }
-            return -1;
-        }
-        return 1;
-    }
-}
