+++ dataformat-core/src/test/java/org/eclipse/digitaltwin/aas4j/v3/dataformat/core/EnumDeserializerTest.java
@@ -1,7 +1,7 @@
 package org.eclipse.digitaltwin.aas4j.v3.dataformat.core;
 
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.databind.DeserializationContext;
+import java.io.IOException;
+
 import org.eclipse.digitaltwin.aas4j.v3.dataformat.core.deserialization.EnumDeserializer;
 import org.eclipse.digitaltwin.aas4j.v3.model.DataTypeIec61360;
 import org.eclipse.digitaltwin.aas4j.v3.model.Direction;
@@ -11,7 +11,8 @@ import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
 
-import java.io.IOException;
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.databind.DeserializationContext;
 
 public class EnumDeserializerTest {
 
@@ -23,6 +24,22 @@ public class EnumDeserializerTest {
         jsonParserMock = Mockito.mock(JsonParser.class);
         deserializationContextMock = Mockito.mock(DeserializationContext.class);
     }
+
+	@Test
+	public void whenDeserializingEnumNames_withUpperCamelCase_shouldReturnScreamingSnakeCase() {
+		Assert.assertEquals("ANY_ENUM", EnumDeserializer.deserializeEnumName("AnyEnum"));
+	}
+
+	@Test
+	public void whenDeserializingEnumNames_withLowerCamelCase_shouldReturnScreamingSnakeCase() {
+		Assert.assertEquals("ANY_ENUM", EnumDeserializer.deserializeEnumName("anyEnum"));
+	}
+
+	@Test
+	public void whenDeserializingEnumNames_withScreamingSnakeCase_shouldReturnUnchanged() {
+		final String name = "ANY_ENUM";
+		Assert.assertEquals(name, EnumDeserializer.deserializeEnumName(name));
+	}
 
     @Test
     public void whenSerializingEnum_usingDataTypeIEC61360_shouldReturnUpperCase() {
@@ -54,7 +71,8 @@ public class EnumDeserializerTest {
         assertDeserialization("off", StateOfEvent.OFF);
     }
 
-    private void assertDeserialization(String value, Enum expected) {
+	@SuppressWarnings("rawtypes")
+	private void assertDeserialization(String value, Enum<?> expected) {
         try {
             Mockito.doReturn(value).when(jsonParserMock).getText();
             Class<? extends Enum> type = expected.getClass();
